- Untested changes to the FS use of the DLM:  - if we already have a lock that covers the extent, use it  - if we need a PR lock and already have a PW lock, use the PW lock  - if we can promote a PR lock to a PW lock and be covered, do the conversion - keep track of the readers and writers using the lock - fixed a lock conversion bug (we were ignoring locks held by other members of   the conversion list when checking for granted compat) - Brought today's changes onto the branch
Next up: fixing the dlm locking mess.

+++ lustre/ldlm/ldlm_resource.c
@@ -263,7 +263,9 @@ void ldlm_resource_dump(struct ldlm_resource *res)
                 LBUG();
 
         snprintf(name, sizeof(name), "%Lx %Lx %Lx",
-                 res->lr_name[0], res->lr_name[1], res->lr_name[2]);
+                 (unsigned long long)res->lr_name[0],
+                 (unsigned long long)res->lr_name[1],
+                 (unsigned long long)res->lr_name[2]);
 
         CDEBUG(D_OTHER, "--- Resource: %p (%s)\n", res, name);
         CDEBUG(D_OTHER, "Namespace: %p (%u)\n", res->lr_namespace,
